user notification policy
