+++ engine/common/api_helpers/custom_fields.py
@@ -1,3 +1,5 @@
+from datetime import timedelta
+
 from django.core.exceptions import ObjectDoesNotExist
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import fields, serializers
@@ -205,3 +207,20 @@ class TimeZoneAwareDatetimeField(serializers.DateTimeField):
             ],
             **kwargs,
         )
+
+
+# TODO: FloatField is used for backward-compatibility, change to IntegerField in a future release
+class DurationSecondsField(serializers.FloatField):
+    def to_internal_value(self, data):
+        return timedelta(seconds=int(super().to_internal_value(data)))
+
+    def to_representation(self, value):
+        return int(value.total_seconds())
+
+
+class DurationMinutesField(serializers.IntegerField):
+    def to_internal_value(self, data):
+        return timedelta(minutes=int(super().to_internal_value(data)))
+
+    def to_representation(self, value):
+        return value.total_seconds() // 60
